migrate allocate action pojo/xcontent to xpack.protocol
This splits out the AllocateAction into some parts xpack-protocol
so that it can be re-usable in the client.
