+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/indexlifecycle/AllocateAction.java
@@ -7,76 +7,21 @@ package org.elasticsearch.xpack.core.indexlifecycle;
 
 import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.metadata.IndexMetaData;
-import org.elasticsearch.common.ParseField;
-import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.xcontent.ConstructingObjectParser;
-import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.xpack.core.indexlifecycle.Step.StepKey;
 
 import java.io.IOException;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 
-public class AllocateAction implements LifecycleAction {
-
-    public static final String NAME = "allocate";
-    public static final ParseField NUMBER_OF_REPLICAS_FIELD = new ParseField("number_of_replicas");
-    public static final ParseField INCLUDE_FIELD = new ParseField("include");
-    public static final ParseField EXCLUDE_FIELD = new ParseField("exclude");
-    public static final ParseField REQUIRE_FIELD = new ParseField("require");
-
-    @SuppressWarnings("unchecked")
-    private static final ConstructingObjectParser<AllocateAction, Void> PARSER = new ConstructingObjectParser<>(NAME,
-            a -> new AllocateAction((Integer) a[0], (Map<String, String>) a[1], (Map<String, String>) a[2], (Map<String, String>) a[3]));
-
-    static {
-        PARSER.declareInt(ConstructingObjectParser.optionalConstructorArg(), NUMBER_OF_REPLICAS_FIELD);
-        PARSER.declareObject(ConstructingObjectParser.optionalConstructorArg(), (p, c) -> p.mapStrings(), INCLUDE_FIELD);
-        PARSER.declareObject(ConstructingObjectParser.optionalConstructorArg(), (p, c) -> p.mapStrings(), EXCLUDE_FIELD);
-        PARSER.declareObject(ConstructingObjectParser.optionalConstructorArg(), (p, c) -> p.mapStrings(), REQUIRE_FIELD);
-    }
-
-    private final Integer numberOfReplicas;
-    private final Map<String, String> include;
-    private final Map<String, String> exclude;
-    private final Map<String, String> require;
-
-    public static AllocateAction parse(XContentParser parser) {
-        return PARSER.apply(parser, null);
-    }
+public class AllocateAction extends org.elasticsearch.protocol.xpack.indexlifecycle.AllocateAction implements LifecycleAction {
 
     public AllocateAction(Integer numberOfReplicas, Map<String, String> include, Map<String, String> exclude, Map<String, String> require) {
-        if (include == null) {
-            this.include = Collections.emptyMap();
-        } else {
-            this.include = include;
-        }
-        if (exclude == null) {
-            this.exclude = Collections.emptyMap();
-        } else {
-            this.exclude = exclude;
-        }
-        if (require == null) {
-            this.require = Collections.emptyMap();
-        } else {
-            this.require = require;
-        }
-        if (this.include.isEmpty() && this.exclude.isEmpty() && this.require.isEmpty() && numberOfReplicas == null) {
-            throw new IllegalArgumentException(
-                    "At least one of " + INCLUDE_FIELD.getPreferredName() + ", " + EXCLUDE_FIELD.getPreferredName() + " or "
-                            + REQUIRE_FIELD.getPreferredName() + "must contain attributes for action " + NAME);
-        }
-        if (numberOfReplicas != null && numberOfReplicas < 0) {
-            throw new IllegalArgumentException("[" + NUMBER_OF_REPLICAS_FIELD.getPreferredName() + "] must be >= 0");
-        }
-        this.numberOfReplicas = numberOfReplicas;
+        super(numberOfReplicas, include, exclude, require);
     }
 
     @SuppressWarnings("unchecked")
@@ -85,46 +30,24 @@ public class AllocateAction implements LifecycleAction {
                 (Map<String, String>) in.readGenericValue());
     }
 
-    public Integer getNumberOfReplicas() {
-        return numberOfReplicas;
-    }
-
-    public Map<String, String> getInclude() {
-        return include;
-    }
-
-    public Map<String, String> getExclude() {
-        return exclude;
-    }
-
-    public Map<String, String> getRequire() {
-        return require;
+    public static AllocateAction parse(XContentParser parser) {
+        org.elasticsearch.protocol.xpack.indexlifecycle.AllocateAction clientAction =
+            org.elasticsearch.protocol.xpack.indexlifecycle.AllocateAction.parse(parser);
+        return new AllocateAction(clientAction.getNumberOfReplicas(), clientAction.getInclude(), clientAction.getExclude(),
+            clientAction.getRequire());
     }
 
     @Override
     public void writeTo(StreamOutput out) throws IOException {
-        out.writeOptionalVInt(numberOfReplicas);
-        out.writeGenericValue(include);
-        out.writeGenericValue(exclude);
-        out.writeGenericValue(require);
+        out.writeOptionalVInt(getNumberOfReplicas());
+        out.writeGenericValue(getInclude());
+        out.writeGenericValue(getExclude());
+        out.writeGenericValue(getRequire());
     }
 
     @Override
     public String getWriteableName() {
         return NAME;
-    }
-
-    @Override
-    public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-        builder.startObject();
-        if (numberOfReplicas != null) {
-            builder.field(NUMBER_OF_REPLICAS_FIELD.getPreferredName(), numberOfReplicas);
-        }
-        builder.field(INCLUDE_FIELD.getPreferredName(), include);
-        builder.field(EXCLUDE_FIELD.getPreferredName(), exclude);
-        builder.field(REQUIRE_FIELD.getPreferredName(), require);
-        builder.endObject();
-        return builder;
     }
 
     @Override
@@ -138,12 +61,12 @@ public class AllocateAction implements LifecycleAction {
         StepKey allocationRoutedKey = new StepKey(phase, NAME, AllocationRoutedStep.NAME);
 
         Settings.Builder newSettings = Settings.builder();
-        if (numberOfReplicas != null) {
-            newSettings.put(IndexMetaData.SETTING_NUMBER_OF_REPLICAS, numberOfReplicas);
+        if (getNumberOfReplicas() != null) {
+            newSettings.put(IndexMetaData.SETTING_NUMBER_OF_REPLICAS, getNumberOfReplicas());
         }
-        include.forEach((key, value) -> newSettings.put(IndexMetaData.INDEX_ROUTING_INCLUDE_GROUP_SETTING.getKey() + key, value));
-        exclude.forEach((key, value) -> newSettings.put(IndexMetaData.INDEX_ROUTING_EXCLUDE_GROUP_SETTING.getKey() + key, value));
-        require.forEach((key, value) -> newSettings.put(IndexMetaData.INDEX_ROUTING_REQUIRE_GROUP_SETTING.getKey() + key, value));
+        getInclude().forEach((key, value) -> newSettings.put(IndexMetaData.INDEX_ROUTING_INCLUDE_GROUP_SETTING.getKey() + key, value));
+        getExclude().forEach((key, value) -> newSettings.put(IndexMetaData.INDEX_ROUTING_EXCLUDE_GROUP_SETTING.getKey() + key, value));
+        getRequire().forEach((key, value) -> newSettings.put(IndexMetaData.INDEX_ROUTING_REQUIRE_GROUP_SETTING.getKey() + key, value));
         UpdateSettingsStep allocateStep = new UpdateSettingsStep(allocateKey, allocationRoutedKey, client, newSettings.build());
         AllocationRoutedStep routedCheckStep = new AllocationRoutedStep(allocationRoutedKey, nextStepKey, true);
         return Arrays.asList(allocateStep, routedCheckStep);
@@ -154,30 +77,5 @@ public class AllocateAction implements LifecycleAction {
         StepKey allocateKey = new StepKey(phase, NAME, NAME);
         StepKey allocationRoutedKey = new StepKey(phase, NAME, AllocationRoutedStep.NAME);
         return Arrays.asList(allocateKey, allocationRoutedKey);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(numberOfReplicas, include, exclude, require);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == null) {
-            return false;
-        }
-        if (obj.getClass() != getClass()) {
-            return false;
-        }
-        AllocateAction other = (AllocateAction) obj;
-        return Objects.equals(numberOfReplicas, other.numberOfReplicas) &&
-            Objects.equals(include, other.include) &&
-            Objects.equals(exclude, other.exclude) &&
-            Objects.equals(require, other.require);
-    }
-
-    @Override
-    public String toString() {
-        return Strings.toString(this);
     }
 }
